feat(mixed-precision): #36 fp32-factor + FP64-IR cuDSS solve + probe knob
The 8j mechanism wired end-to-end:
- device_csr_fp32(): torch fp32 CSR over the round-on-store snapshot for the
  cuDSS fp32 factorization (stable-operand #37 contract, fp32; flat + #38
  chunked concat).
- solvers/iterative_refinement.fp64_iterative_refinement: backend-agnostic
  FP64-IR loop — residual r=b-Ax via an injected FP64 matvec (against the
  fp64 vals_d, NEVER the fp32 factor), correction via the fp32 factor_solve;
  converge to 1e-12 rel or 10 sweeps; returns (x, refinement-count) (§8j).
- wodo_film._solve_device -> _solve_device_fp32_ir and exciton
  _cudss_dev_solve -> _cudss_dev_solve_fp32_ir: fp32 cuDSS factor + FP64-IR,
  fp64 SpMV via device_operator(); >10 sweeps = conditioning FAIL (G2, NaN /
  ConvergenceError to the reject ladder); per-solve count logged.
- val_dtype plumbed through XDDSystem/XDDDeviceAssembler and the film's
  _val_dtype class-attr idiom (like _index_width/_chunking).
- gh200_capacity_probe --val-dtype fp32 (mirrors --chunking/--force-narrow;
  help notes fp32-factor+IR needs linsolver=cudss).

Gates (CPU, Mac): the FP64-IR loop logic (recovers fp64 from an fp32 factor
in a few sweeps; residual uses the fp64 matvec not the fp32 factor; >max_iter
flagged non-converged; zero-rhs exact), film/XDD default-fp64-unchanged,
probe arg wiring. 13 passed. cuDSS-backed G1/G3/G4/G5 are CUDA-only (gpubox /
Nova lanes).

+++ src/diffsim/physics/exciton_device.py
@@ -577,13 +577,14 @@ class XDDDeviceAssembler:
     fields are scaled by diag(n̂)/diag(p̂).
     """
 
-    def __init__(self, sysm, index_width="auto"):
+    def __init__(self, sysm, index_width="auto", val_dtype="fp64"):
         dm = sysm.dm
         self.sysm = sysm
         self.dm = dm
         self.asm = DeviceNSAssembler(dm, ndof=NDOF, node_pattern=True,
                                      blockmask=XDD_BLOCKMASK,
-                                     index_width=index_width)
+                                     index_width=index_width,
+                                     val_dtype=val_dtype)
         d = dm.device
         self._dev = d
